+++ src/ui/help.rs
@@ -1,7 +1,18 @@
 pub fn get_help_docs() -> Vec<Vec<&'static str>> {
     vec![
         vec!["General", "a", "Jump to currently playing album"],
+        vec![
+            "General",
+            "A",
+            "Jump to currently playing artist's album list",
+        ],
         vec!["General", "s", "Save track"],
+        vec!["General", "+", "Increase volume by 10%"],
+        vec!["General", "-", "Decrease volume by 10%"],
+        vec!["General", "n", "Skip to next track"],
+        vec!["General", "p", "Skip to previous track"],
+        vec!["General", "<", "Seek backwards 5 seconds"],
+        vec!["General", ">", "Seek forwards 5 seconds"],
         vec!["General", "<Ctrl+s>", "Toggle shuffle"],
         vec!["General", "<Ctrl+r>", "Cycle repeat mode"],
         vec!["General", "h | <Left Arrow Key>", "Move selection left"],
@@ -21,7 +32,7 @@ pub fn get_help_docs() -> Vec<Vec<&'static str>> {
         vec!["General", "<Enter>", "Enter active mode"],
         vec![
             "General",
-            "q | -",
+            "q",
             "Go back or exit when nowhere left to back to",
         ],
         vec!["General", "d", "Select device to play music on"],